+++ unit-tests/json-to-iac/power-vs-instances.test.js
@@ -47,6 +47,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -93,6 +96,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_pin_policy        = "soft"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -147,6 +153,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -196,6 +205,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_type      = "tier1"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -311,6 +323,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -372,6 +387,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_type      = "tier1"
   pi_network {
     network_id = ERROR: Unfound Ref
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -488,6 +506,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -603,6 +624,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_type      = "tier1"
   pi_network {
     network_id = data.ibm_pi_network.power_network_example_dev_nw.id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -721,6 +745,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = data.ibm_pi_network.power_network_example_dev_nw.id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -771,6 +798,9 @@ data
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -815,6 +845,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_sap_profile_id    = "ush1-4x256"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -864,6 +897,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
     ip_address = "1.2.3.4"
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -909,6 +945,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_pool      = "Tier1-Flash2"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -960,6 +999,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_type          = "tier1"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -1018,6 +1060,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -1073,6 +1118,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_affinity_volume       = ibm_pi_volume.example_volume_oracle_1_db_1.volume_id
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -1131,6 +1179,9 @@ resource "ibm_pi_instance" "example_workspace_instance_frog" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 `;
       assert.deepEqual(
@@ -1179,6 +1230,9 @@ resource "ibm_pi_instance" "example_falconstor_vtl_test" {
   pi_license_repository_capacity = 1
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 `;
@@ -1237,6 +1291,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 
 ##############################################################################
@@ -1294,6 +1351,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_storage_type      = "tier1"
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }
 
@@ -1373,6 +1433,9 @@ resource "ibm_pi_instance" "example_workspace_instance_test" {
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
   }
+  timeouts {
+    create = "3h"
+  }
 }
 
 ##############################################################################
@@ -1397,6 +1460,9 @@ resource "ibm_pi_instance" "example_falconstor_vtl_test" {
   pi_license_repository_capacity = 1
   pi_network {
     network_id = ibm_pi_network.power_network_example_dev_nw.network_id
+  }
+  timeouts {
+    create = "3h"
   }
 }